+++ README.md
@@ -1,4 +1,11 @@
-# KekUpload server
+```
+██╗  ██╗███████╗██╗  ██╗██╗   ██╗██████╗ ██╗      ██████╗  █████╗ ██████╗ 
+██║ ██╔╝██╔════╝██║ ██╔╝██║   ██║██╔══██╗██║     ██╔═══██╗██╔══██╗██╔══██╗
+█████╔╝ █████╗  █████╔╝ ██║   ██║██████╔╝██║     ██║   ██║███████║██║  ██║
+██╔═██╗ ██╔══╝  ██╔═██╗ ██║   ██║██╔═══╝ ██║     ██║   ██║██╔══██║██║  ██║
+██║  ██╗███████╗██║  ██╗╚██████╔╝██║     ███████╗╚██████╔╝██║  ██║██████╔╝
+╚═╝  ╚═╝╚══════╝╚═╝  ╚═╝ ╚═════╝ ╚═╝     ╚══════╝ ╚═════╝ ╚═╝  ╚═╝╚═════╝ 
+```
 
 ![](https://tokei.rs/b1/github/KotwOSS/kekupload-server)
 ![](https://tokei.rs/b1/github/KotwOSS/kekupload-server?category=blanks)
