+++ build.rs
@@ -2,7 +2,10 @@
 use std::process::Command;
 fn main() {
     // note: add error checking yourself.
-    let output = Command::new("git").args(&["rev-parse", "--short", "HEAD"]).output().unwrap();
+    let output = Command::new("git")
+        .args(&["rev-parse", "--short", "HEAD"])
+        .output()
+        .unwrap();
     let git_hash = String::from_utf8(output.stdout).unwrap();
     println!("cargo:rustc-env=GIT_HASH={}", git_hash);
 }
