+++ src/config.rs
@@ -1,3 +1,9 @@
+/*
+* Created on Wed Jun 01 2022
+*
+* Copyright (c) 2022 KotwOSS
+*/
+
 pub const EXTENSION_MAX_LENGTH: usize = 10;
 pub const STREAM_ID_LENGTH: usize = 64;
 pub const FILE_ID_LENGTH: usize = 7;
