+++ src/errors.rs
@@ -1,3 +1,9 @@
+/*
+* Created on Wed Jun 01 2022
+*
+* Copyright (c) 2022 KotwOSS
+*/
+
 use actix_web::{
     error,
     http::{header::ContentType, StatusCode},
@@ -91,7 +97,7 @@ json_error_type!(MISSING, StatusCode::BAD_REQUEST);
 json_error_type!(DB_CONNECT, StatusCode::INTERNAL_SERVER_ERROR);
 json_error_type!(DB_QUERY, StatusCode::INTERNAL_SERVER_ERROR);
 
-json_error_type!(FS_CREATE, StatusCode::INTERNAL_SERVER_ERROR); 
+json_error_type!(FS_CREATE, StatusCode::INTERNAL_SERVER_ERROR);
 json_error_type!(FS_RENAME, StatusCode::INTERNAL_SERVER_ERROR);
 json_error_type!(FS_WRITE, StatusCode::INTERNAL_SERVER_ERROR);
 json_error_type!(FS_OPEN, StatusCode::INTERNAL_SERVER_ERROR);
