+++ src/main.rs
@@ -1,3 +1,9 @@
+/*
+* Created on Wed Jun 01 2022
+*
+* Copyright (c) 2022 KotwOSS
+*/
+
 #[macro_use]
 extern crate diesel;
 
@@ -31,21 +37,38 @@ fn clean_tmp<'a>(tmp: &'a str) {
 }
 
 fn print_logo() {
-    println!("{}\n\
+    println!(
+        "{}\n\
  █▄▀ █▀▀ █▄▀ █ █ █▀█ █   █▀█ ▄▀█ █▀▄\n\
  █ █ ██▄ █ █ █▄█ █▀▀ █▄▄ █▄█ █▀█ █▄▀\n\
-    {}", 
-    colors::ORANGE, colors::RESET);
-    println!("KekUpload version {}{}{}\n\n", colors::LIGHT_BLUE, env!("GIT_HASH"), colors::RESET);
-    println!("Written by {}KotwOSS{} licensed under {}MIT{}.\n",
-        colors::LIGHT_BLUE, colors::RESET, colors::LIGHT_BLUE, colors::RESET);
-    println!("{}(c){} Copyright 2022 {}KotwOSS{}\n\n", 
-        colors::RED, colors::RESET, colors::LIGHT_BLUE, colors::RESET);
+    {}",
+        colors::ORANGE,
+        colors::RESET
+    );
+    println!(
+        "KekUpload version {}{}{}\n\n",
+        colors::LIGHT_BLUE,
+        env!("GIT_HASH"),
+        colors::RESET
+    );
+    println!(
+        "Written by {}KotwOSS{} licensed under {}MIT{}.\n",
+        colors::LIGHT_BLUE,
+        colors::RESET,
+        colors::LIGHT_BLUE,
+        colors::RESET
+    );
+    println!(
+        "{}(c){} Copyright 2022 {}KotwOSS{}\n\n",
+        colors::RED,
+        colors::RESET,
+        colors::LIGHT_BLUE,
+        colors::RESET
+    );
 }
 
 #[tokio::main]
 async fn main() {
-
     print_logo();
 
     dotenv().ok();
