--- src/routes/api/mod.rs
@@ -1,6 +1,0 @@
-pub mod create;
-pub mod download;
-pub mod finish;
-pub mod remove;
-pub mod upload;
-pub mod length;
