+++ src/routes/mod.rs
@@ -1,3 +1,9 @@
+/*
+* Created on Wed Jun 01 2022
+*
+* Copyright (c) 2022 KotwOSS
+*/
+
 pub mod api;
 pub mod embed;
 pub mod web;
