+++ src/util/mod.rs
@@ -1,4 +1,10 @@
+/*
+* Created on Wed Jun 01 2022
+*
+* Copyright (c) 2022 KotwOSS
+*/
+
 pub mod checker;
+pub mod chunked;
 pub mod files;
 pub mod random;
-pub mod chunked;
